+++ wiki/Changelog.md
@@ -1,3 +1,4 @@
+  * Enabled character attributes italic (#418, #152) and strikeout.
   * True Colour support (#431) (using 38;2;r;g;b m) alpha.
   * Added MSYS setup hint (#426).
   * Fixed CSI 8;...t handling of default and zero values (issue #408).